+++ src/main/java/com/besson/arknights/block/custom/lgd/DuplexWoodenFence.java
@@ -12,14 +12,10 @@ public class DuplexWoodenFence extends FurnitureHorizontalFacingBlock {
     public DuplexWoodenFence(Settings settings) {
         super(settings);
     }
-    private static final VoxelShape SHAPE_NS = Block.createCuboidShape(0, 0, 7, 16, 16, 9);
-    private static final VoxelShape SHAPE_WE = Block.createCuboidShape(7, 0, 0, 9, 16, 16);
+    private static final VoxelShape SHAPE = Block.createCuboidShape(0, 0, 0, 16, 16, 16);
 
     @Override
     public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
-        return switch (state.get(FACING)){
-            case WEST,EAST -> SHAPE_WE;
-            default -> SHAPE_NS;
-        };
+        return SHAPE;
     }
 }